Replaced obsolete API usages as of DNN 8.0.4
Replaced obsolete API usages as of DNN 8.0.4
The module currently targets DNN 8.0.4 and was already using obsolete APIs in the current state targetting this version (sometimes deprecated in DNN6!!!)

This refactors those usages to use the APIs available in v8.0.4. After this PR is merged, I'll start updating DNN dependencies.